requirements.txt
alembic==1.4.2
appdirs==1.4.3
blinker==1.4
click==7.1.2
cycler==0.10.0
distlib==0.3.0
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.17
virtualenv==20.0.15
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
